Split rest interface tests into multiple files.
As suggested by Dan in a review.
